% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\title{Bubble evidence map plotting function}
\usage{
bubbleplot(
  x_var,
  xlabels = "",
  y_var,
  ylabels = "",
  clr_var,
  n,
  wrap_width = 10,
  bg_col = "#BFD5E3",
  title = "",
  subtitle = "",
  x_title = "",
  y_title = "",
  palette = "Set2",
  interactive = FALSE,
  hovertext = ""
)
}
\arguments{
\item{x_var}{Variable to plot on the x-axis. Variable should be ordinal
or categorical. Vectors of strings are converted to ordinal for plotting,
with original levels shown on the plot.}

\item{xlabels}{Optional vector of strings used to plot x-axis labels if
x_var is ordinal.}

\item{y_var}{Variable to plot on the y-axis. Variable should be ordinal
or categorical. Vectors of strings are converted to ordinal for plotting,
with original levels shown on the plot.}

\item{ylabels}{Optional vector of strings used to plot x-axis labels if
y_var is ordinal.}

\item{clr_var}{Variable (vector) used to colour bubbles. Up to five
levels of the vector may be plotted as different bubbles.}

\item{n}{Numerical variable corresponding to the number of studies and
used to alter the size of the bubbles.}

\item{wrap_width}{The width of axis labels using character wrapping.
The default is 'wrap_width = 10'.}

\item{bg_col}{The background colour for the plot area. The default is
'bg_col = "#BFD5E3"'.}

\item{title}{Optional plot title.}

\item{subtitle}{Optional plot subtitle.}

\item{x_title}{X-axis title.}

\item{y_title}{Y-axis title.}

\item{palette}{Palette colour used to colour bubbles. The default is
'palette = "Set2"'.}

\item{interactive}{Logical argument (TRUE or FALSE) specifying whether the
plot should be made interactive or not. Interactive plots include a
hoverover tooltip containing additional text.}

\item{hovertext}{Optional string containing text to plot in the hoverover
tooltip.}
}
\value{
Plots an interactive bubble evidence map (using ggplotly)
}
\description{
Plots a bubble evidence map for a systematic map, evidence
gap map or similar.
}
\examples{
\dontrun{
data <- read.csv2('inst/extdata/data.csv', sep=',')
attach(data)
plot <- bubbleplot(x_var = inst, 
    y_var = outcome, 
    clr_var = loc,
    n = n,
    bg_col = '#F1EFF3',
    title = 'RCTs in Education',
    subtitle = 'Data from 1980 to 2017',
    x_title = 'Institution',
    y_title = 'Outcome',
    palette = 'Set2',
    interactive = TRUE,
    hovertext = 'test')
plot
}
}
