% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_interactive.R
\name{go_interactive}
\alias{go_interactive}
\title{Interactive bubble and grid evidence map}
\usage{
go_interactive(plot, type, input1, input2)
}
\arguments{
\item{plot}{A ggplot2 object produced from bubble().}

\item{type}{The type of plot, either 'bubble' or 'grid' (heatmap).}

\item{input1}{If a bubble plot is provided, this should correspond to
the dataframe containing the full plot data. If a grid or heatmap is
provided, this should correspond to a vector of URLs, one for each
plotted datapoint, listed in the same order as the data provided in
the plotted dataframe.}

\item{input2}{If a bubble plot is provided, this should correspond to
the clr_var categorical variable used to colour bubbles. If a grid or
heatmap is plotted, this should be left empty.}
}
\value{
An interactive plot object.
}
\description{
Convert a static bubble or grid plot to an interactive,
HTML version with embedded hyperlinks.
}
\examples{
\dontrun{
go_interactive(plot, type = 'grid', data$url)
go_interactive(plot, type = 'bubble', input1 = data, input2 = loc)
}
}
