% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summdat.R
\name{summdat}
\alias{summdat}
\title{Summarise dataframe by counting number of rows}
\usage{
summdat(data, col1, col2, col3)
}
\arguments{
\item{data}{Dataframe containing individuals rows corresponding to studies in a
systematic map, for example.}

\item{col1}{String corresponding to the column name of one of up to three
factors (categorical variables) for which counts will be made of the number of
rows corresponding to each level of the factor.}

\item{col2}{String corresponding to the column name of one of up to three
factors (categorical variables) for which counts will be made of the number of
rows corresponding to each level of the factor.}

\item{col3}{String corresponding to the column name of one of up to three
factors (categorical variables) for which counts will be made of the number of
rows corresponding to each level of the factor.}
}
\value{
A dataframe containing the summary counts of the number of rows across
each level of each of up to three variables.
}
\description{
Counts the number of rows across up to three variables of interest.
}
\examples{
col1 <- 'outcome'
col2 <- 'inst'
col3 <- 'loc'
text <- RCurl::getURL("https://raw.githubusercontent.com/Cuan-crypto/EGM_RCT/main/EDRCT.csv")
data <- read.csv(text = text)
summary <- summdat(data, col1, col2, col3)
summary;
}
